refactor: add shared folder in api
