+++ src/features/RedactForm/hooks/useRedactForm.ts
@@ -2,7 +2,7 @@ import { useAppDispatch } from '@/shared/hooks/useAppDispatch'
 import { useNavigate } from 'react-router-dom'
 import { RedactSchema, type Redact } from '../consts/RedactSchema'
 import { useEffect } from 'react'
-import { refreshUser } from '@/api/refreshUser'
+import { refreshUser } from '@/api/shared/refreshUser'
 import { useForm, type UseFormReturn } from 'react-hook-form'
 import useLoading from '@/shared/hooks/useLoading'
 import { getRoute } from '@/shared/helpers/getRoute'
