+++ src/shared/consts/categories.ts
@@ -1,4 +1,4 @@
-import { getCategories } from '@/api/getCategories'
+import { getCategories } from '@/api/shared/getCategories'
 import { processingRequestResult } from '@/shared/helpers/processingRequestHandlers/processingRequestResult'
 
 export const categories = processingRequestResult(await getCategories()) || []
